// web/src/api/models/RepositoryBookmark.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepositoryBookmark
 */
export interface RepositoryBookmark {
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    createdDatetime?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    folderId?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    icon?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    link?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    updatedDatetime?: string;
    /**
     * 
     * @type {string}
     * @memberof RepositoryBookmark
     */
    userId?: string;
}

/**
 * Check if a given object implements the RepositoryBookmark interface.
 */
export function instanceOfRepositoryBookmark(value: object): value is RepositoryBookmark {
    return true;
}

export function RepositoryBookmarkFromJSON(json: any): RepositoryBookmark {
    return RepositoryBookmarkFromJSONTyped(json, false);
}

export function RepositoryBookmarkFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepositoryBookmark {
    if (json == null) {
        return json;
    }
    return {
        
        'createdDatetime': json['created_datetime'] == null ? undefined : json['created_datetime'],
        'description': json['description'] == null ? undefined : json['description'],
        'folderId': json['folder_id'] == null ? undefined : json['folder_id'],
        'icon': json['icon'] == null ? undefined : json['icon'],
        'id': json['id'] == null ? undefined : json['id'],
        'link': json['link'] == null ? undefined : json['link'],
        'name': json['name'] == null ? undefined : json['name'],
        'updatedDatetime': json['updated_datetime'] == null ? undefined : json['updated_datetime'],
        'userId': json['user_id'] == null ? undefined : json['user_id'],
    };
}

export function RepositoryBookmarkToJSON(json: any): RepositoryBookmark {
    return RepositoryBookmarkToJSONTyped(json, false);
}

export function RepositoryBookmarkToJSONTyped(value?: RepositoryBookmark | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'created_datetime': value['createdDatetime'],
        'description': value['description'],
        'folder_id': value['folderId'],
        'icon': value['icon'],
        'id': value['id'],
        'link': value['link'],
        'name': value['name'],
        'updated_datetime': value['updatedDatetime'],
        'user_id': value['userId'],
    };
}